refactor(stella-mcp): split toolset.rs under the file-size ceiling after the #1875/#1898 growth
toolset.rs reached 1756 lines after three stacked merges — #1875 (sort the
MCP schema segment for byte-stability, #1848), #1898 (bound one server's
contribution to the advertised toolset, #1856), and the parked-waits work
(#1860) — crossing the 1500-line hard limit for files not in
scripts/file-size-baseline.txt and turning main's file-size gate red.

Move the #[cfg(test)] tests module out to src/toolset/tests.rs as a
sibling submodule (the driver/settlement.rs pattern; same shape as
stella-model's zai/tests.rs and anthropic/tests.rs). Purely mechanical:
the block is de-indented one level and declared as `mod tests;` — no
visibility changes (a child module sees the parent's private items via
`use super::*`), no renames, no behavior change. toolset.rs lands at 812
lines, toolset/tests.rs at 950. No baseline entry, per the guard's own
policy: split, do not grandfather.
